Move several tests to SharedWithSparse
